+++ box2d/src/main/kotlin/ktx/box2d/worlds.kt
@@ -152,3 +152,36 @@ fun World.rayCast(
     callback: KtxRayCastCallback) {
   rayCast(callback, startX, startY, endX, endY)
 }
+
+/**
+ * Query the world for all fixtures that potentially overlap the provided AABB (Axis-Aligned Bounding Box).
+ *
+ * @param lowerX the x coordinate of the lower left corner
+ * @param lowerY the y coordinate of the lower left corner
+ * @param upperX the x coordinate of the upper right corner
+ * @param upperY the y coordinate of the upper right corner
+ * @param callback a user implemented callback that is called for every fixture overlapping the AABB.
+ * @see Query
+ */
+fun World.query(lowerX: Float, lowerY: Float, upperX: Float, upperY: Float, callback: KtyQueryCallback) {
+  QueryAABB(callback, lowerX, lowerY, upperX, upperY)
+}
+
+/**
+ * Stores constants that can be returned by [KtyQueryCallback] to control its behavior.
+ * @see query
+ */
+object Query {
+  /**
+   * Stop querying the world.
+   * @see KtyQueryCallback
+   */
+  const val STOP = false
+  /**
+   * Continue querying for the next match.
+   * @see KtyQueryCallback
+   */
+  const val CONTINUE = true
+}
+
+typealias KtyQueryCallback = (Fixture) -> Boolean